Revert unrelated changes
