--- .github/workflows/pull_request.yml
@@ -1,31 +1,0 @@
-name: "pull_request"
-
-on: pull_request
-
-jobs:
-  check:
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
-
-      - name: Set up Java
-        uses: actions/setup-java@v3
-        with:
-          distribution: 'zulu'
-          java-version: 17
-
-      - name: check
-        uses: gradle/gradle-build-action@v2
-        with:
-          arguments: build
-
-      - name: Upload reports
-        if: failure()
-        uses: actions/upload-artifact@v3
-        with:
-          name: 'reports'
-          path: '**/build/reports/**'
