+++ sample/build.gradle.kts
@@ -19,8 +19,5 @@ tasks.findByName("rewriteRun")!!.dependsOn(rootProject.tasks.jar)
 tasks.findByName("rewriteDiscover")!!.dependsOn(rootProject.tasks.jar)
 
 rewrite {
-    activeRecipe(
-        "arrow.RaiseRefactor",
-        "arrow.RaiseRecipe"
-      )
+    activeRecipe("arrow.RaiseRefactor")
 }
