+++ settings.gradle.kts
@@ -1,2 +1,3 @@
 rootProject.name = "rewrite-arrow"
+
 include(":sample")
