--- src/test/java/arrow/EagerEffectScopeTest.java
@@ -1,105 +1,0 @@
-package arrow;
-
-import org.junit.jupiter.api.Test;
-import org.openrewrite.config.Environment;
-import org.openrewrite.kotlin.KotlinParser;
-import org.openrewrite.test.RecipeSpec;
-import org.openrewrite.test.RewriteTest;
-
-import static org.openrewrite.kotlin.Assertions.kotlin;
-
-class EagerEffectScopeTest implements RewriteTest {
-    @Override
-    public void defaults(RecipeSpec spec) {
-        spec.recipe(
-            Environment.builder()
-              .scanRuntimeClasspath()
-              .build()
-              .activateRecipes(
-                "arrow.RaiseRefactor",
-                "arrow.RaiseRecipe"
-              )
-          )
-          .parser(
-            KotlinParser.builder()
-              .logCompilationWarningsAndErrors(true)
-              .classpath("arrow-core-jvm")
-          );
-    }
-
-    @Test
-    void eagerEffectScopeParameter() {
-        rewriteRun(
-          kotlin(
-            """
-              package com.yourorg
-                            
-              import arrow.core.continuations.EagerEffectScope
-
-              fun test(scope: EagerEffectScope<String>): Int {
-                return scope.shift("failure")
-              }
-              """,
-            """
-              package com.yourorg
-                            
-              import arrow.core.raise.Raise
-
-              fun test(scope: Raise<String>): Int {
-                return scope.shift("failure")
-              }
-              """
-          )
-        );
-    }
-
-    @Test
-    void eagerEffectScopeReceiver() {
-        rewriteRun(
-          kotlin(
-            """
-              package com.yourorg
-                            
-              import arrow.core.continuations.EagerEffectScope
-
-              fun EagerEffectScope<String>.test(): Int {
-                return shift("failure")
-              }
-              """,
-            """
-              package com.yourorg
-                            
-              import arrow.core.raise.Raise
-
-              fun Raise<String>.test(): Int {
-                return shift("failure")
-              }
-              """
-          )
-        );
-    }
-
-    @Test
-    void eagerEffectScopeReceiverExpression() {
-        rewriteRun(
-          kotlin(
-            """
-              package com.yourorg
-                            
-              import arrow.core.continuations.EagerEffectScope
-
-              fun EagerEffectScope<String>.test(): Int =
-                shift("failure")
-              """,
-            """
-              package com.yourorg
-                            
-              import arrow.core.raise.Raise
-
-              fun Raise<String>.test(): Int =
-                shift("failure")
-              """
-          )
-        );
-    }
-}
