+++ src/test/java/arrow/RaiseRecipeTest.java
@@ -8,17 +8,14 @@ import org.openrewrite.test.RewriteTest;
 
 import static org.openrewrite.kotlin.Assertions.kotlin;
 
-class EffectScopeTest implements RewriteTest {
+class RaiseRecipeTest implements RewriteTest {
     @Override
     public void defaults(RecipeSpec spec) {
         spec.recipe(
             Environment.builder()
               .scanRuntimeClasspath()
               .build()
-              .activateRecipes(
-                "arrow.RaiseRecipe",
-                "arrow.RaiseRefactor"
-              )
+              .activateRecipes("arrow.RaiseRefactor")
           )
           .parser(
             KotlinParser.builder()
@@ -123,6 +120,82 @@ class EffectScopeTest implements RewriteTest {
 
               suspend fun Raise<Int>.test(): Unit =
                 ensure(false) { -1 }
+              """
+          )
+        );
+    }
+
+    @Test
+    void eagerEffectScopeParameter() {
+        rewriteRun(
+          kotlin(
+            """
+              package com.yourorg
+                            
+              import arrow.core.continuations.EagerEffectScope
+
+              fun test(scope: EagerEffectScope<String>): Int {
+                return scope.shift("failure")
+              }
+              """,
+            """
+              package com.yourorg
+                            
+              import arrow.core.raise.Raise
+
+              fun test(scope: Raise<String>): Int {
+                return scope.shift("failure")
+              }
+              """
+          )
+        );
+    }
+
+    @Test
+    void eagerEffectScopeReceiver() {
+        rewriteRun(
+          kotlin(
+            """
+              package com.yourorg
+                            
+              import arrow.core.continuations.EagerEffectScope
+
+              fun EagerEffectScope<String>.test(): Int {
+                return shift("failure")
+              }
+              """,
+            """
+              package com.yourorg
+                            
+              import arrow.core.raise.Raise
+
+              fun Raise<String>.test(): Int {
+                return shift("failure")
+              }
+              """
+          )
+        );
+    }
+
+    @Test
+    void eagerEffectScopeReceiverExpression() {
+        rewriteRun(
+          kotlin(
+            """
+              package com.yourorg
+                            
+              import arrow.core.continuations.EagerEffectScope
+
+              fun EagerEffectScope<String>.test(): Int =
+                shift("failure")
+              """,
+            """
+              package com.yourorg
+                            
+              import arrow.core.raise.Raise
+
+              fun Raise<String>.test(): Int =
+                shift("failure")
               """
           )
         );
